fix: eliminate cyclic imports — move _filter_entity_aliases to catalog_merger
Moves _filter_entity_aliases() into catalog_merger.py so it can be used
without importing from semantic_extraction, breaking the import cycle.
semantic_extraction.py now imports it from catalog_merger instead.
